// interactiveslidesaddinui/src/enums/figures.ts
export enum Figures {
    MULTIPLE_CHOICE = 'multiple-choice',
    RANGE_SELECTOR = 'range-selector',
    OPEN_FIELD = 'open-field',
    CUSTOMER_DETAILS = 'customer-details',
    BUTTON = 'button',
    CALCULATOR = 'calculator',
    DATE_FIELD = 'date-field',
    PERSONALIZATION = 'personalization'
}